pyproject.toml
[tool.poetry]
name = "zenodotus"
version = "0.1.0"
description = ""
authors = ["Joe Kaufeld <joe.kaufeld@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
django-reversion = "^3.0.9"
djangorestframework = "^3.12.2"
django-taggit = "^1.3.0"
Pillow = "^8.1.0"
django-unused-media = "^0.2.2"
gunicorn = "^20.0.4"
black = "^20.8b1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
